Cleanup smelly code
