+++ Cartfile.resolved
@@ -1,1 +1,1 @@
-github "adyen/adyen-3ds2-ios" "2.1.0-rc.5"
+github "adyen/adyen-3ds2-ios" "2.1.0-rc.6"
